auto-claims-demo/repair-shop-agent/pyproject.toml
[project]
name = "auto-claims-repair-shop-agent"
version = "0.1.0"
description = "An AI agent for finding and communicating with repair shops for the Auto Claims Demo."
authors = [
    {name = "Your Name", email = "your@email.com"},
]
dependencies = [
    "google-adk[bigquery-analytics]>=1.29.0",
    "a2a-sdk~=0.3.24",
    "nest-asyncio>=1.6.0,<2.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "gcsfs>=2024.11.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-aiplatform[agent-engines]>=1.148.0",
    "protobuf>=6.31.1,<7.0.0",
    "opentelemetry-instrumentation-httpx>=0.59b0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-grpc>=0.59b0",
    "ruff>=0.15.10",
    "authlib==1.6.11",
    "mako==1.3.11",
    "python-multipart==0.0.26",
]
requires-python = ">=3.10,<3.14"


[dependency-groups]
dev = [
    "pytest>=9.0.3,<10.0.0",
    "pytest-asyncio>=0.23.8,<2.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
]

[project.optional-dependencies]
jupyter = [
    "jupyter>=1.0.0,<2.0.0",
]

lint = [
    "ruff>=0.4.6,<1.0.0",
    "ty>=0.0.1a0",
    "codespell>=2.2.0,<3.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "B006"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["app", "frontend"]

[tool.ty]
# ty is Astral's Rust-based type checker (same team as ruff/uv)
# See: https://docs.astral.sh/ty/

[tool.ty.environment]
python-version = "3.10"

[tool.ty.src]
exclude = [".venv/**"]

[tool.ty.rules]
# Ignore common issues with third-party libraries and dynamic code patterns
unresolved-import = "ignore"
unresolved-attribute = "ignore"
invalid-argument-type = "ignore"
invalid-assignment = "ignore"
invalid-return-type = "ignore"
possibly-missing-attribute = "ignore"
not-subscriptable = "ignore"
deprecated = "ignore"

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb,**/package-lock.json"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["app","frontend"]

[tool.agent-starter-pack]
# Generation metadata - enables CLI commands to understand project context
# These fields allow any generated project to work as a remote template
name = "auto-claims-repair-shop-agent"
description = "An AI agent for finding and communicating with repair shops for the Auto Claims Demo."
base_template = "adk_a2a"
agent_directory = "app"

# Generation context
generated_at = "2026-02-22T18:37:36.834642+00:00"
asp_version = "0.38.0"

[tool.agent-starter-pack.create_params]
# Parameters used during project creation
deployment_target = "agent_engine"
session_type = "in_memory"
cicd_runner = "google_cloud_build"
include_data_ingestion = false
datastore = "none"
